% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AllScaledData}
\alias{AllScaledData}
\title{Scaled abundance data for 2 subsets of 4 rays in the Irish Sea, by gbm.cons}
\format{A data frame with 378570 rows and 3 variables}
\usage{
data(AllScaledData)
}
\description{
A dataset containing the output of the gbm.cons example run, conservation
priority areas within the Irish Sea for juvenile and adult female cuckoo,
blonde, thornback and spotted rays.
}
\details{
\itemize{
  \item Longitude. Decimal longitudes in the Irish Sea
  \item Latitude. Decimal latitudes in the Irish Sea
  \item allscaled. Relative abundance. Each juvenile and adult female
  cuckoo, blonde, thornback and spotted ray scaled to 1 and added together.
}
}
\keyword{datasets}

