% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration}
\alias{calibration}
\title{calibration}
\usage{
calibration(obs, preds, family = c("binomial", "bernoulli", "poisson"))
}
\arguments{
\item{obs}{Observed data.}

\item{preds}{Predicted data.}

\item{family}{Statistical distribution family. Choose one.}
}
\value{
roc & calibration stats internally within gbm runs e.g. in gbm.auto.
}
\description{
Internal use only. Jane Elith/John Leathwick 17th March 2005. Calculates
calibration statistics for either binomial or count data but the family
argument must be specified for the latter a conditional test for the latter
will catch most failures to specify the family.
}
\author{
Simon Dedman, \email{simondedman@gmail.com}
}
