% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.auto.R
\name{gbm.auto}
\alias{gbm.auto}
\title{Automated Boosted Regression Tree modelling and mapping suite}
\usage{
gbm.auto(grids = NULL, samples = mysamples, expvar, resvar, lr = c(0.01,
  0.005), bf = 0.5, ZI = "CHECK", simp = TRUE, gridslat = 2,
  gridslon = 1, cols = grey.colors(1, 1, 1), linesfiles = FALSE,
  savegbm = TRUE, varint = TRUE, map = TRUE, RSB = TRUE, BnW = TRUE,
  alerts = TRUE, pngtype = "cairo-png", binonly = FALSE, ...)
}
\arguments{
\item{grids}{Explantory data to predict to. Import with (e.g.) read.csv & specify object name. Defaults to NULL (won't predict to grids)}

\item{samples}{Explanatory & response variables to predict from. Keep col names short, no odd characters, starting numerals or terminal periods. Spaces may be converted to periods in directory names, underscores won't. Can be a subset. Default is mysamples}

\item{expvar}{List of column numbers of explanatory variables in 'samples', expected e.g. c(1,35,67,etc.). No default}

\item{resvar}{Column number of response variable (e.g. CPUE) in samples. Expected, e.g. 12. No default. Column name should be species name. tc = c(2,5),         # list of permutations of tree complexity allowed, defaults to c(2,length(expvar)) @L55}

\item{lr}{List of permutations of learning rate allowed}

\item{bf}{Bag fraction permutations, can be vector: c(0.5,0.7)}

\item{ZI}{are data zero-inflated? TRUE/FALSE/"CHECK". TRUE: delta BRT, log-normalised Gaus, reverse log-norm & bias corrected. FALSE: do Gaussian only, no log-normalisation. CHECK: Tests data for you. Default is TRUE.}

\item{simp}{Try simplfying best BRTs?}

\item{gridslat}{Column number for latitude in 'grids'}

\item{gridslon}{Column number for longitude in 'grids'}

\item{cols}{Barplot colour vector. Assignment in order of explanatory variables. Default 1*white: white bars black borders. '1*' repeats}

\item{linesfiles}{Save individual line plots' data as csv's?}

\item{savegbm}{Save gbm objects? Open w/ load("Bin_Best_Model")}

\item{varint}{Calculate variable interactions? Default:TRUE, FALSE for error "contrasts can be applied only to factors with 2 or more levels"}

\item{map}{Save abundance map png files?}

\item{RSB}{Run Unrepresentativeness surface builder?}

\item{BnW}{Repeat maps in black & white e.g. for print journals}

\item{alerts}{Play sounds to mark progress steps}

\item{pngtype}{Filetype for png files, alternatively try "quartz"}

\item{...}{Optional arguments for tc, gbm.map & gbm.step}

\item{mapshape}{Set coast shapefile, else downloaded & autogenerated}
}
\value{
Line, dot & bar plots, a report of all variables used, statistics for
tests, variable interactions, predictors used & dropped, etc. If selected,
generates predicted abundance maps, and Unrepresentativeness surface
}
\description{
Automates delta log normal boosted regression trees abundance prediction.
Loops through all permutations of parameters provided (learning
rate, tree complexity, bag fraction), chooses the best, then simplifies it.
Generates line, dot & bar plots, and outputs these and the predictions
and a report of all variables used, statistics for tests, variable
interactions, predictors used & dropped, etc.. If selected, generates
predicted abundance maps, and Unrepresentativeness surfaces.
}
\details{
Errors and their origins:
1. Error in FUN(X[[i]], ...) : only defined on a data frame with all numeric variables
> Explanatory variables are expected to be numeric
2. At bf=0.5, if nrows <= 42 gbm.step will crash
> Use gbm.bfcheck to determine optimal viable bf size
3. Maps/plots dont work/output
> If on a Mac, try changing pngtype to "quartz"
4. Error in while (delta.deviance > tolerance.test & n.fitted < max.trees)  :
 missing value where TRUE/FALSE needed
 > Data are expected to contain zeroes (lots of them in zero-inflated cases),
 have you already filtered them out?
}
\examples{
gbm.auto(expvar = c(4:7, 9, 11), resvar = 12, grids = mygrids,
lr = c(0.005, 0.001), ZI = TRUE, savegbm = FALSE, mapshape = coast)

}
\author{
Simon Dedman, \email{simondedman@gmail.com}
}

