% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.predict.grids.R
\name{gbm.predict.grids}
\alias{gbm.predict.grids}
\title{gbm.predict.grids}
\usage{
gbm.predict.grids(
  model,
  new.dat,
  want.grids = F,
  preds2R = T,
  sp.name = "preds",
  pred.vec = NULL,
  filepath = NULL,
  num.col = NULL,
  num.row = NULL,
  xll = NULL,
  yll = NULL,
  cell.size = NULL,
  no.data = NULL,
  plot = F,
  full.grid = T,
  part.number = NULL,
  part.row = NULL,
  header = T
)
}
\arguments{
\item{model}{Gbm model object.}

\item{new.dat}{New data to predict to. Must be a data frame with column names
identical to names for all variables in the model used for prediction.}

\item{want.grids}{If used, grids are written to a nominated directory and
optionally also plotted in R.}

\item{preds2R}{Write predictions to R directory.}

\item{sp.name}{Species prediction name}

\item{pred.vec}{A vector of -9999's, the length of the scanned full grid
(i.e. without nodata values excluded).}

\item{filepath}{Must specify the whole path as a character vector,but without
the final file name - eg "c:/gbm/".}

\item{num.col}{Number of columns for grids.}

\item{num.row}{Number of rows for grids.}

\item{xll}{X latitude longitude corner.}

\item{yll}{Y latitude longitude corner.}

\item{cell.size}{Cell size.}

\item{no.data}{No data value.}

\item{plot}{Plot output.}

\item{full.grid}{Predict to full grid else to part grid.}

\item{part.number}{Number of grid subset part.}

\item{part.row}{Number of rows of grid subset part.}

\item{header}{Want.grids header included.}
}
\value{
Powers the predictive mapping element of gbm.map.
}
\description{
Internal use only. J.Elith / J.Leathwick, March 07. To make predictions to
sites or grids. If to sites, the predictions are written to the R workspace.
If to grid, the grids are written to a nominated directory and optionally
also plotted in R. New data (new.dat) must be a data frame with column names
identical to names for all variables in the model used for prediction.
pred.vec is a vector of -9999's, the length of the scanned full grid (i.e.
without nodata values excluded).filepath must specify the whole path as a
character vector,but without the final file name - eg "c:/gbm/"
}
\examples{
None
}
\author{
Simon Dedman, \email{simondedman@gmail.com}

Jane Elith

John Leathwick
}
