% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.subset.R
\name{gbm.subset}
\alias{gbm.subset}
\title{Subset data into two groups}
\usage{
gbm.subset(x, fams = c("Bin", "Gaus"), loop = FALSE)
}
\arguments{
\item{x}{Vector of variable names.}

\item{fams}{statistical data distribution family names to be modelled by gbm.}

\item{loop}{Is the folder a gbm.loop output?}
}
\description{
Returns the variable value corresponding to the 0 value on the lineplots,
which should be the optimal place to split the dataset into 2 subsets.
}
\details{
loop varnames are BinLineLoop_VAR.csv & GausLineLoop_VAR.csv
 normal varnames are Bin_Best_line_VAR.csv & Gaus_Best_line_VAR.csv

 just use average between the last negative & first positive point
 unless any points fall on zero
}
\examples{
None

}
\author{
Simon Dedman, \email{simondedman@gmail.com}
}
