% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grids}
\alias{grids}
\title{Explanatory variables for rays in the Irish Sea}
\format{A data frame with 378570 rows and 43 variables}
\usage{
data(grids)
}
\description{
A dataset containing explanatory variables for environment, fishery and
predators of rays including juveniles in the Irish Sea.
}
\details{
\itemize{
  \item Longitude. Decimal longitudes in the Irish Sea
  \item Latitude. Decimal latitudes in the Irish Sea
  \item Depth. Metres, decimal.
  \item Temperature. Degrees, decimal.
  \item Salinity. PPM.
  \item Current_Speed. Metres per second at the seabed.
  \item Distance_to_Shore. Metres, decimal.
  \item F_LPUE. Commercial fishery LPUE in Kg/Hr
  \item Scallop.
  \item MI_Av_E_Hr.
  \item MI_Av_LPUE.
  \item MI_Sum_Liv.
  \item Whelk.
  \item MmoAvScKwh.
  \item HubDist. Distance from nearest Voronoi polygon centroid
  \item Cod_C. ICES IBTS CPUE of cod caught between 1994 - 2014 large enough to predate upon <= year 1 cuckoo rays.
  \item Cod_T. As Cod_C for yr1 thornback rays.
  \item Cod_B. As Cod_C for yr1 blonde rays.
  \item Cod_S. As Cod_C for yr1 spotted rays.
  \item Haddock_C. As Cod_C, haddock predating upon cuckoo rays
  \item Haddock_T. As Cod_C, haddock predating upon thornback rays
  \item Haddock_B. As Cod_C, haddock predating upon blonde rays
  \item Haddock_S. As Cod_C, haddock predating upon spotted rays
  \item Plaice_C. As Cod_C, plaice predating upon cuckoo rays
  \item Plaice_T. As Cod_C, plaice predating upon thornback rays
  \item Plaice_B. As Cod_C, plaice predating upon blonde rays
  \item Plaice_S. As Cod_C, plaice predating upon spotted rays
  \item Whiting_C. As Cod_C, whiting predating upon cuckoo rays
  \item Whiting_T. As Cod_C, whiting predating upon thornback rays
  \item Whiting_B. As Cod_C, whiting predating upon blonde rays
  \item Whiting_S. As Cod_C, whiting predating upon spotted rays
  \item ComSkt_C. As Cod_C, common skate predating upon cuckoo rays
  \item ComSkt_T. As Cod_C, common skate predating upon thornback rays
  \item ComSkt_B. As Cod_C, common skate predating upon blonde rays
  \item ComSkt_S. As Cod_C, common skate predating upon spotted rays
  \item Blonde_C. As Cod_C, blonde ray predating upon cuckoo rays
  \item Blonde_T. As Cod_C, blonde ray predating upon thornback rays
  \item Blonde_S. As Cod_C, blonde ray predating upon spotted rays
  \item C_Preds. All predator CPUEs combined for cuckoo rays.
  \item T_Preds. All predator CPUEs combined for thornback rays.
  \item B_Preds. All predator CPUEs combined for blonde rays.
  \item S_Preds. All predator CPUEs combined for spotted rays.
  \item Effort.
}
}
\keyword{datasets}

