% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples}
\alias{samples}
\title{Numbers of 4 ray species caught in 2137 Irish Sea trawls, 1994 to 2014}
\format{A data frame with 2244 rows and 14 variables}
\usage{
data(samples)
}
\description{
2244 capture events of cuckoo, thornback, spotted and blonde rays in the Irish
Sea from 1994 to 2014 by the ICES IBTS, including explanatory variables:
Length Per Unit Effort in that area by the commercial fishery, fishing effort
by same, depth, temperature, distance to shore, and current speed at the
bottom. Note column 8 between Distance_to_shore and Current_Speed is blank.
}
\details{
\itemize{
  \item Survey_StNo_HaulNo_Year.
  \item Latitude. Decimal latitudes in the Irish Sea
  \item Longitude. Decimal longitudes in the Irish Sea
  \item Depth. Metres, decimal.
  \item Temperature. Degrees, decimal.
  \item Salinity. PPM.
  \item Current_Speed. Metres per second at the seabed.
  \item Distance_to_Shore. Metres, decimal.
  \item F_LPUE. Commercial fishery LPUE in Kg/Hr
  \item Effort.
  \item Cuckoo. Numbers of juvenile cuckoo rays caught, standardised to 1 hour
  \item Thornback. Numbers of juvenile thornback rays caught, standardised to 1 hour
  \item Blonde. Numbers of juvenile blonde rays caught, standardised to 1 hour
  \item Spotted. Numbers of juvenile spotted rays caught, standardised to 1 hour
}
}
\keyword{datasets}

