% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.auto.R
\name{gbm.auto}
\alias{gbm.auto}
\title{Automated Boosted Regression Tree modelling and mapping suite}
\usage{
gbm.auto(grids = NULL, samples, expvar, resvar, tc = c(2), lr = c(0.01,
  0.005), bf = 0.5, ZI = "CHECK", fam1 = "bernoulli", fam2 = "gaussian",
  simp = TRUE, gridslat = 2, gridslon = 1, cols = grey.colors(1, 1, 1),
  linesfiles = TRUE, smooth = FALSE, savegbm = TRUE, varint = TRUE,
  map = TRUE, shape = NULL, RSB = TRUE, BnW = TRUE, alerts = TRUE,
  pngtype = "cairo-png", gaus = TRUE, ...)
}
\arguments{
\item{grids}{Explantory data to predict to. Import with (e.g.) read.csv and
specify object name. Defaults to NULL (won't predict to grids)}

\item{samples}{Explanatory and response variables to predict from. Keep col
names short, no odd characters, starting numerals or terminal periods. Spaces
 may be converted to periods in directory names, underscores won't. Can be a
 subset}

\item{expvar}{List of column numbers of explanatory variables in 'samples',
expected e.g. c(1,35,67,etc.). No default}

\item{resvar}{Column number(s) of response variable (e.g. CPUE) in samples.
e.g. 12 or c(4,5,6). No default. Column name should be species name}

\item{tc}{Permutations of tree complexity allowed, can be vector with
the largest sized number no larger than the number of explanatory variables
e.g. c(2,7), or a list of 2 single numbers or vectors, the first to be passed
to the binary BRT, the second to the Gaussian, e.g. tc = list(c(2,6), 2) or
list(6, c(2,6))}

\item{lr}{Permutations of learning rate allowed. Can be a vector or a list of
2 single numbers or vectors, the first to be passed to the binary BRT, the
second to the Gaussian, e.g. lr = list(c(0.01,0.02),0.0001) or
list(0.01,c(0.001, 0.0005))}

\item{bf}{Permutations of bag fraction allowed, can be single number, vector
or list, per tc and lr}

\item{ZI}{are data zero-inflated? TRUE FALSE "CHECK". TRUE: delta BRT,
log-normalised Gaus, reverse log-norm and bias corrected. FALSE: do Gaussian
only, no log-normalisation. CHECK: Tests data for you. Default is TRUE.}

\item{fam1}{probability distribution family for 1st part of delta process,
defaults to "bernoulli"}

\item{fam2}{probability distribution family for 2nd part of delta process,
defaults to "gaussian"}

\item{simp}{Try simplfying best BRTs?}

\item{gridslat}{Column number for latitude in 'grids'}

\item{gridslon}{Column number for longitude in 'grids'}

\item{cols}{Barplot colour vector. Assignment in order of explanatory
variables. Default 1*white: white bars black borders. '1*' repeats}

\item{linesfiles}{Save individual line plots' data as csv's?}

\item{smooth}{Apply a smoother to the line plots? Default FALSE}

\item{savegbm}{Save gbm objects and make available in environment after
running? Open with load("Bin_Best_Model")}

\item{varint}{Calculate variable interactions? Default:TRUE, FALSE for error
"contrasts can be applied only to factors with 2 or more levels"}

\item{map}{Save abundance map png files?}

\item{shape}{Set coast shapefile, else bounds calculated by gbm.map which
then calls gbm.basemap to download and autogenerate the base map.}

\item{RSB}{Run Unrepresentativeness surface builder?}

\item{BnW}{Repeat maps in black and white e.g. for print journals?}

\item{alerts}{Play sounds to mark progress steps?}

\item{pngtype}{Filetype for png files, alternatively try "quartz"}

\item{gaus}{Do Gaussian (fam2) runs as well as Bin? Default TRUE.}

\item{...}{Optional arguments for zero in breaks.grid in gbm.map, legend in
legend.grid in gbm.map, and gbm.step (dismo)}
}
\value{
Line, dot and bar plots, a report of all variables used, statistics
for tests, variable interactions, predictors used and dropped, etc. If
selected generates predicted abundance maps, and Unrepresentativeness surface
}
\description{
Automates delta log normal boosted regression trees abundance prediction.
Loops through all permutations of parameters provided (learning
rate, tree complexity, bag fraction), chooses the best, then simplifies it.
Generates line, dot and bar plots, and outputs these and the predictions
and a report of all variables used, statistics for tests, variable
interactions, predictors used and dropped, etc.. If selected, generates
predicted abundance maps, and Unrepresentativeness surfaces.
}
\details{
Errors and their origins:

1. Error in FUN(X[[i]], ...) : only defined on a data frame with all numeric variables
> Explanatory variables are expected to be numeric

2. At bf=0.5, if nrows <= 42 gbm.step will crash
> Use gbm.bfcheck to determine optimal viable bf size

3. Maps/plots dont work/output
> If on a Mac, try changing pngtype to "quartz"

4. Error in while (delta.deviance > tolerance.test AMPERSAND n.fitted < max.trees)  :
 missing value where TRUE/FALSE needed
> Data are expected to contain zeroes (lots of them in zero-inflated cases),
have you already filtered them out?

5. Error in round(gbm.object$cv.statistics$deviance.mean, 4) : non-numeric
argument to mathematical function
> LR or BF probably too low in earlier BRT (normally Gaus run with highest TC)

6. Error in if (n.trees > x$n.trees) { : argument is of length zero}
> LR or BF probably too low in earlier BRT (normally Gaus run with highest TC)

7. Error in gbm.fit(x, y, offset = offset, distribution = distribution, w = w)
The dataset size is too small or subsampling rate is too large:
nTrain*bag.fraction <= n.minobsinnode
> LR or BF probably too low in earlier BRT (normally Gaus run with highest TC)
It may be that you don't have enough positive samples to run BRT modelling
Run gbm.bfcheck to check recommended minimum BF size

8. Warning message: In cor(y_i, u_i) : the standard deviation is zero
> LR or BF probably too low in earlier BRT (normally Gaus run with highest TC)
It may be that you don't have enough positive samples to run BRT modelling
Run gbm.bfcheck to check recommended minimum BF size

9. Anomalous values can obfuscate clarity in line plots e.g. salinity range
32:35 degrees but dataset has errant 0 value: plot axis 0:35, 99.99% data
in the tiny bit at the right. Clean your data well beforehand.
}
\examples{
gbm.auto(expvar = c(4:8, 10), resvar = 11, grids = mygrids,
tc = c(2,7), lr = c(0.005, 0.001), ZI = TRUE, savegbm = FALSE)

}
\author{
Simon Dedman, \email{simondedman@gmail.com}
}
