% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.basemap.R
\name{gbm.basemap}
\alias{gbm.basemap}
\title{Creates Basemaps for Gbm.auto mapping from your data range}
\usage{
gbm.basemap(bounds, getzip = TRUE, zipvers = "2.3.5-1",
  savename = "Crop_Map", res = "CALC")
}
\arguments{
\item{bounds}{Region to crop to: c(xmin,xmax,ymin,ymax)}

\item{getzip}{Download & unpack GSHHS data to WD? "TRUE" else absolute/relative reference to GSHHS_shp folder, inc that folder}

\item{zipvers}{GSHHS version, in case it updates. Please email developer (SD) if this is incorrect}

\item{savename}{Shapefile savename, no extension}

\item{res}{resolution, 1:5 (low:high) OR c,l,i,h,f (coarse, low, intermediate, high, full) or "CALC" to calculate based on bounds}
}
\value{
basemap coastline file for gbm.map in gbm.auto
}
\description{
Downloads unzips crops & saves NOAAs global coastline shapefiles to user-set
box. Use for 'shape' in gbm.map. If downloading in RStudio uncheck
"Use secure download method for HTTP" in Tools > Global Options > Packages.
Simon Dedman, 2015/6 simondedman@gmail.com github.com/SimonDedman/gbm.auto
}
\examples{

}

