% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.cons.R
\name{gbm.cons}
\alias{gbm.cons}
\title{Conservation Area Mapping}
\usage{
gbm.cons(mygrids, subsets, conssamples, alerts = TRUE, map = TRUE,
  BnW = TRUE, resvars, gbmautos = TRUE, expvars, tcs = NULL,
  lrs = rep(list(c(0.01, 0.005)), length(resvars)), bfs = rep(0.5,
  length(resvars)), ZIs = rep("CHECK", length(resvars)), gridslats = rep(2,
  length(resvars)), gridslons = rep(1, length(resvars)),
  colss = rep(list(grey.colors(1, 1, 1)), length(resvars)),
  linesfiless = rep(FALSE, length(resvars)), savegbms = rep(TRUE,
  length(resvars)), varints = rep(TRUE, length(resvars)), maps = rep(TRUE,
  length(resvars)), RSBs = rep(TRUE, length(resvars)), BnWs = rep(TRUE,
  length(resvars)), zeroes = rep(TRUE, length(resvars)),
  pngtype = "cairo-png")
}
\arguments{
\item{mygrids}{Gridded lat+long+data object to predict to}

\item{subsets}{Subset name(s): character; single or vector}

\item{conssamples}{Single/vector of samples csv names & locations corresponding to subsets}

\item{alerts}{Play sounds to mark progress steps}

\item{map}{Produce maps}

\item{BnW}{Also produce B&W maps?}

\item{resvars}{Vector of resvars cols from conssamples for gbm.autos, length(subsets)*species, no default}

\item{gbmautos}{Do gbm.auto runs for species?}

\item{expvars}{List object of expvar vectors for gbm.autos, length = no. of subsets * no. of species. No default}

\item{tcs}{Gbm.auto paramaters, autocalculated below if not provided by user}

\item{lrs}{Gbm.auto paramaters, autocalculated below if not provided by user}

\item{bfs}{Gbm.auto paramaters, autocalculated below if not provided by user}

\item{ZIs}{Gbm.auto paramaters, autocalculated below if not provided by user}

\item{gridslats}{Gbm.auto paramaters, autocalculated below if not provided by user}

\item{gridslons}{Gbm.auto paramaters, autocalculated below if not provided by user}

\item{colss}{Gbm.auto paramaters, autocalculated below if not provided by user}

\item{linesfiless}{Gbm.auto paramaters, autocalculated below if not provided by user}

\item{savegbms}{Gbm.auto paramaters, autocalculated below if not provided by user}

\item{varints}{Gbm.auto paramaters, autocalculated below if not provided by user}

\item{maps}{Gbm.auto paramaters, autocalculated below if not provided by user}

\item{RSBs}{Gbm.auto paramaters, autocalculated below if not provided by user}

\item{BnWs}{Gbm.auto paramaters, autocalculated below if not provided by user}

\item{zeroes}{Gbm.auto paramaters, autocalculated below if not provided by user}

\item{pngtype}{Filetype for png files, alternatively try "quartz"}
}
\value{
Maps via gbm.map & saved data as csv file
}
\description{
Runs gbm.auto for multiple subsets of the same overall dataset and
scales the combined results, leading to maps which highlight areas of high
conservation importance for multiple species in the same study area e.g.
using juvenile and adult female subsets to locate candidate nursery grounds
and spawning areas respectively.
}
\examples{

}
\author{
Simon Dedman, \email{simondedman@gmail.com}
}

