% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.loop.R
\name{gbm.loop}
\alias{gbm.loop}
\title{Calculate Coefficient Of Variation surfaces for gbm.auto predictions}
\usage{
gbm.loop(loops = 10, savecsv = TRUE, calcpreds = TRUE, varmap = T,
  measure = "CPUE", cleanup = FALSE, grids = NULL, samples, expvar,
  resvar, tc = c(2), lr = c(0.01), bf = 0.5, ZI = "CHECK",
  simp = TRUE, gridslat = 2, gridslon = 1, cols = grey.colors(1, 1, 1),
  linesfiles = F, savegbm = F, varint = F, map = F, shape = NULL,
  RSB = F, BnW = F, alerts = T, pngtype = "cairo-png", ...)
}
\arguments{
\item{loops}{The number of loops required, integer}

\item{savecsv}{Save the coefficients of variation in simple & extended format}

\item{calcpreds}{Calculate coefficients of variation of predicted abundance?}

\item{varmap}{Create a map of the coefficients of variation outputs?}

\item{measure}{Map legend, coefficients of variation of what? Default CPUE}

\item{cleanup}{Remove gbm.auto-generated directory each loop? Default FALSE}

\item{grids}{See gbm.auto help for all subsequent params}

\item{samples}{See gbm.auto help}

\item{expvar}{See gbm.auto help}

\item{resvar}{See gbm.auto help}

\item{tc}{See gbm.auto help}

\item{lr}{See gbm.auto help}

\item{bf}{See gbm.auto help}

\item{ZI}{See gbm.auto help}

\item{simp}{See gbm.auto help}

\item{gridslat}{See gbm.auto help}

\item{gridslon}{See gbm.auto help}

\item{cols}{See gbm.auto help}

\item{linesfiles}{See gbm.auto help}

\item{savegbm}{See gbm.auto help}

\item{varint}{See gbm.auto help}

\item{map}{See gbm.auto help}

\item{shape}{See gbm.auto help}

\item{RSB}{See gbm.auto help}

\item{BnW}{See gbm.auto help}

\item{alerts}{See gbm.auto help}

\item{pngtype}{See gbm.auto help}

\item{...}{Additional params for gbm.auto subfunctions inc gbm.step}
}
\value{
Returns a data frame of lat, long, 1 predicted abundance per loop,
and a final variance score per cell.
}
\description{
Processes a user-specified number of loops through the same gbm.auto
parameter combinations and calculates the Coefficient Of Variation in the
predicted abundance scores for each site aka cell. This can be mapped to
spatially demonstrate the output variance range.
}
\examples{
library("gbm.auto")
mygrids <- gbm.auto::grids # load grids
mysamples <- gbm.auto::samples # load samples
gbmloopexample <- gbm.loop(loops = 2, samples = mysamples,
grids = mygrids, expvar = c(4:10), resvar = 11, simp = F)

}
\author{
Simon Dedman, \email{simondedman@gmail.com}
}
