% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.map.R
\name{gbm.map}
\alias{gbm.map}
\title{Maps of predicted abundance from Boosted Regression Tree modelling}
\usage{
gbm.map(x, y, z, byx = NULL, byy = NULL, grdfun = mean,
  mapmain = "Predicted CPUE (numbers per hour): ",
  species = "Response Variable", heatcolours = c("white", "yellow",
  "orange", "red", "brown4"), colournumber = 8, shape = coast,
  landcol = "grey80", mapback = "lightblue", legendloc = "bottomright",
  legendtitle = "CPUE", lejback = "white", zero = TRUE, quantile = 1,
  byxout = FALSE, breaks = NULL, ...)
}
\arguments{
\item{x}{Vector of longitudes, from make.grid in mapplots; x. Order by this (descending) SECOND}

\item{y}{Vector of latitudes, from make.grid in mapplots; grids[,gridslat]. ##Order by this (descending) first}

\item{z}{Vector of abundances generated by gbm.predict.grids, from make.grid in mapplots; grids[,predabund]}

\item{byx}{Longitudinal width of grid cell, from make.grid in mapplots. Autogenerated if left blank.}

\item{byy}{Latitudinal height of grid cell, from make.grid in mapplots. Autogenerated if left blank.}

\item{grdfun}{make.grid operand for >=2 values per cell. Default:mean, other options: sum prod min max sd se var}

\item{mapmain}{Output map image response variable title, from basemap in mapplots}

\item{species}{Response variable name, from basemap in mapplots; names(samples[i]). Defaults to "Response Variable"}

\item{heatcolours}{Abundance colour scale, defaults to the heatcol from legend.grid & draw.grid in mapplots.}

\item{colournumber}{number of colours to spread heatcol over, default:8}

\item{shape}{Number of colours to spread heatcol over, default:8}

\item{landcol}{Colour for 'null' area of map, if appropriate, from draw.shape in mapplots. Was "darkgreen" changed to light grey}

\item{mapback}{Basemap background colour}

\item{legendloc}{Location on map of legend box, from legend.grid in mapplots}

\item{legendtitle}{The metric of abundance, e.g. CPUE for fisheries, from legend.grid in mapplots}

\item{lejback}{Background colour of legend, from legend.grid in mapplots}

\item{zero}{Allow 0 category in breaks.grid & thus legend? Default TRUE}

\item{quantile}{Set max breakpoint; lower this to cutoff outliers}

\item{byxout}{Export byx to use elsewhere? Default:FALSE}

\item{breaks}{Vector of breakpoints for colour scales; default blank, generated automatically}

\item{...}{Additional parameters for mapplots functions, especially legend.grid}
}
\value{
Species abundance maps using data provided by gbm.auto, and
Representativeness Surface Builder maps using data provided by gbm.rsb, to be
run in a png/par/gbm.map/dev.off sequence.
}
\description{
Generates maps from the outputs of gbm.step then gbm.predict.grids, handled
automatically within gbm.auto but can be run alone, and generates
representativeness surfaces from the output of gbm.rsb
}
\examples{
Suggested code for outputting to e.g. png:
png(...)
par(...)
gbm.map(...)
dev.off()

Suggested parameter values for use with gbm.rsb:
z = rsbdf[,"Unrepresentativeness"]
mapmain = "Unrepresentativeness: "
legendtitle = "UnRep 0-1"
}

