% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.valuemap.R
\name{gbm.valuemap}
\alias{gbm.valuemap}
\title{Decision Support Tool that generates (Marine) Protected Area options using
species predicted abundance maps}
\usage{
gbm.valuemap(
  dbase,
  loncolno = 1,
  latcolno = 2,
  goodcols,
  badcols,
  conservecol = NULL,
  plotthis = c("good", "bad", "both", "close"),
  maploops = c("Combo", "Biomass", "Effort", "Conservation"),
  savethis = TRUE,
  HRMSY = 0.15,
  goodweight = NULL,
  badweight = NULL,
  m = 1,
  alerts = TRUE,
  BnW = TRUE,
  shape = NULL,
  pngtype = "cairo-png",
  byxport = NULL,
  ...
)
}
\arguments{
\item{dbase}{Data.frame to load. Expects Lon, Lat & data columns: predicted
abundances, fishing effort etc. E.g.: Abundance_Preds_All.csv from gbm.auto.}

\item{loncolno}{Column number in dbase which has longitudes.}

\item{latcolno}{Column number in dbase which has latitudes.}

\item{goodcols}{Which column numbers are abundances (where higher = better)?
List them in order of highest conservation importance first e.g. c(3,1,2,4).}

\item{badcols}{Which col no.s are 'negative' e.g. fishing (where higher =
worse)?}

\item{conservecol}{Conservation column, from gbm.cons.}

\item{plotthis}{To plot? delete any,or all w/ NULL.}

\item{maploops}{Sort loops to run.}

\item{savethis}{Export all data as csv?}

\item{HRMSY}{Maximum percent of each goodcols stock which can be removed
yearly, as decimal (0.15 = 15 pct). Must protect remainder: 1-HRMSY. Single
number or vector. If vector, same order as goodcols. Required.}

\item{goodweight}{Single/vector weighting multiple(s) for goodcols array.}

\item{badweight}{Ditto for badcols array.}

\item{m}{Multiplication factor for Bpa units. 1000 to convert tonnes to
kilos, 0.001 kilos to tonnes. Assumedly the same for all goodcols.}

\item{alerts}{Play sounds to mark progress steps.}

\item{BnW}{Also produce greyscale images for print publications.}

\item{shape}{Set coastline shapefile, else uses British Isles. Generate your
own with gbm.basemap.}

\item{pngtype}{Filetype for png files, alternatively try "quartz".}

\item{byxport}{Dummy param for package testing for CRAN, ignore.}

\item{...}{Optional terms for gbm.map.}
}
\value{
Species abundance, abundance vs avoid variable, and MPA maps per
species and sort type, in b&w if set. CSVs of all maps if set.
}
\description{
Scales response variable data, maps a user-defined explanatory variable to be
avoided, e.g. fishing effort, combines them into a map showing areas to
preferentially close. Bpa, the precautionary biomass required to protect the
spawning stock, is used to calculate MPA size. MPA is then grown to add
subsequent species starting from the most conservationally at-risk species,
resulting in one MPA map per species, and a multicolour MPA map of all.
All maps list the percentage of the avoidvariables total that is overlapped
by the MPA in the map legend.
}
\author{
Simon Dedman, \email{simondedman@gmail.com}
}
